Release 0.2.2 (#65)
* OAK D SR POE and some reliability improvements (#35)

* WIP add support for tof (oak d sr poe)

* Accommodate new sdk, make tof stream show up, more or less properly. Still need to fix a few gui bugs.

* Fixed gui bugs with which sensors are "AI capable" improved a bit on resolution search stability.  did py-lints

* install wheel from artifactory

* Update rerun_py/depthai_viewer/_backend/device.py

* Update rerun_py/depthai_viewer/_backend/device.py

* Use sentry.luxonis

* Added in the new tof_decoding.

* Py lints

* Fixed rebase issues

* Fixed tiny1c temperature resolution. Bumped depthai tof decoding branch. Added new filters. Don't fail when sentry init fails. Likely fix viewer crashing when there is no internet available.

* fix pylints

* OAK-D-LR better default, fixed point cloud coloring when using albedo texture, aligned depth from ToF on oak d sr poe to CAM_B

* Lints

* Added --viewer-mode flag.

* release-0.2.2, added --viewer-mode flag

* Changelog

+++ crates/re_viewer/src/app.rs
@@ -51,6 +51,7 @@ pub struct StartupOptions {
     pub memory_limit: re_memory::MemoryLimit,
     pub persist_state: bool,
     pub sdk_port: u32,
+    pub viewer_mode: bool,
 }
 
 #[derive(Clone, Default)]
@@ -571,37 +572,39 @@ impl eframe::App for App {
                 }
                 dependency_installer.update();
             }
-            match &mut self.backend_handle {
-                Some((handle, port)) => match handle.try_wait() {
-                    Ok(status) => {
-                        match status {
-                            Some(_) => {
-                                let _ = handle.kill(); // It will only Err in case the process is already dead (which is what we want anyway)
-                                self.state.depthai_state.reset();
-                                re_log::debug!("Backend process has exited, restarting!");
-                                self.backend_handle = self.spawn_backend(&self.backend_environment);
-                            }
-                            None => {
-                                if !self.state.depthai_state.backend_comms.ws.is_initialized() {
-                                    self.state.depthai_state.backend_comms.ws.connect(*port);
+            if !self.startup_options.viewer_mode {
+                match &mut self.backend_handle {
+                    Some((handle, port)) => match handle.try_wait() {
+                        Ok(status) => {
+                            match status {
+                                Some(_) => {
+                                    let _ = handle.kill(); // It will only Err in case the process is already dead (which is what we want anyway)
+                                    self.state.depthai_state.reset();
+                                    re_log::debug!("Backend process has exited, restarting!");
+                                    self.backend_handle = self.spawn_backend(&self.backend_environment);
+                                }
+                                None => {
+                                    if !self.state.depthai_state.backend_comms.ws.is_initialized() {
+                                        self.state.depthai_state.backend_comms.ws.connect(*port);
+                                    }
                                 }
                             }
                         }
+                        Err(_) => {
+                        }
+                    },
+                    None => {
+                        if self.backend_environment.are_requirements_installed() {
+                            self.backend_handle = self.spawn_backend(&self.backend_environment);
+                        } else {
+                            re_log::debug!(
+                                "Backend requirements not installed, starting dependency installer!"
+                            );
+                            self.install_dependencies();
+                        }
                     }
-                    Err(_) => {
-                    }
-                },
-                None => {
-                    if self.backend_environment.are_requirements_installed() {
-                        self.backend_handle = self.spawn_backend(&self.backend_environment);
-                    } else {
-                        re_log::debug!(
-                            "Backend requirements not installed, starting dependency installer!"
-                        );
-                        self.install_dependencies();
-                    }
-                }
-            };
+                };
+            }
         }
 
         if self.startup_options.memory_limit.limit.is_none() {
